chore: renamed OneClickPaymentComponent to StoredPaymentMethodComponent
