+++ AdyenTests/Card Tests/StoredPaymentMethodComponentTests.swift
@@ -6,13 +6,14 @@
 
 @testable import AdyenDropIn
 @testable import Adyen
+@testable import AdyenCard
 import XCTest
 
 class StoredPaymentMethodComponentTests: XCTestCase {
     
     func testLocalizationWithCustomTableName() {
         let method = StoredPaymentMethodMock(identifier: "id", supportedShopperInteractions: [.shopperNotPresent], type: "test_type", name: "test_name")
-        let sut = OneClickPaymentComponent(paymentMethod: method)
+        let sut = StoredPaymentMethodComponent(paymentMethod: method)
         let payment = Payment(amount: Payment.Amount(value: 34, currencyCode: "EUR"), countryCode: "DE")
         sut.payment = payment
         sut.localizationParameters = LocalizationParameters(tableName: "AdyenUIHost", keySeparator: nil)
@@ -26,7 +27,7 @@ class StoredPaymentMethodComponentTests: XCTestCase {
     
     func testLocalizationWithCustomKeySeparator() {
         let method = StoredPaymentMethodMock(identifier: "id", supportedShopperInteractions: [.shopperNotPresent], type: "test_type", name: "test_name")
-        let sut = OneClickPaymentComponent(paymentMethod: method)
+        let sut = StoredPaymentMethodComponent(paymentMethod: method)
         let payment = Payment(amount: Payment.Amount(value: 34, currencyCode: "EUR"), countryCode: "DE")
         sut.payment = payment
         sut.localizationParameters = LocalizationParameters(tableName: "AdyenUIHostCustomSeparator", keySeparator: "_")
@@ -36,6 +37,53 @@ class StoredPaymentMethodComponentTests: XCTestCase {
         XCTAssertEqual(viewController?.actions.count, 2)
         XCTAssertEqual(viewController?.actions.first?.title, ADYLocalizedString("adyen_cancelButton", sut.localizationParameters))
         XCTAssertEqual(viewController?.actions.last?.title, ADYLocalizedSubmitButtonTitle(with: payment.amount, sut.localizationParameters))
+    }
+
+    func testUI() {
+        let method = StoredPaymentMethodMock(identifier: "id",
+                                             supportedShopperInteractions: [.shopperPresent],
+                                             type: "type",
+                                             name: "name")
+        let sut = StoredPaymentMethodComponent(paymentMethod: method)
+        sut.clientKey = "client_key"
+
+        let delegate = PaymentComponentDelegateMock()
+
+        let delegateExpectation = expectation(description: "expect delegate to be called.")
+        delegate.onDidSubmit = { data, component in
+            XCTAssertTrue(component === sut)
+            XCTAssertNotNil(data.paymentMethod as? StoredPaymentDetails)
+
+            let details = data.paymentMethod as! StoredPaymentDetails
+            XCTAssertEqual(details.type, "type")
+            XCTAssertEqual(details.storedPaymentMethodIdentifier, "id")
+
+            delegateExpectation.fulfill()
+        }
+        delegate.onDidFail = { _, _ in
+            XCTFail("delegate.didFail() should never be called.")
+        }
+        sut.delegate = delegate
+
+        let payemt = Payment(amount: Payment.Amount(value: 174, currencyCode: "EUR"), countryCode: "NL")
+        sut.payment = payemt
+
+        UIApplication.shared.keyWindow?.rootViewController?.present(sut.viewController, animated: false, completion: nil)
+
+        let uiExpectation = expectation(description: "Dummy Expectation")
+        DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + .seconds(1)) {
+            let alertController = sut.viewController as! UIAlertController
+
+            XCTAssertTrue(alertController.actions.contains { $0.title == ADYLocalizedString("adyen.cancelButton", nil) })
+            XCTAssertTrue(alertController.actions.contains { $0.title == ADYLocalizedSubmitButtonTitle(with: payemt.amount, nil) })
+
+            let payAction = alertController.actions.first { $0.title == ADYLocalizedSubmitButtonTitle(with: payemt.amount, nil) }!
+
+            payAction.tap()
+
+            uiExpectation.fulfill()
+        }
+        waitForExpectations(timeout: 10, handler: nil)
     }
     
 }
